+++ src/cloud_functions/functions/src/index.ts
@@ -1,6 +1,7 @@
 import * as functions from 'firebase-functions';
 import admin = require('firebase-admin');
 import notifications = require('./services/user_notifications');
+import utils = require('./utils');
 
 admin.initializeApp();
 const db = admin.firestore();
@@ -39,7 +40,7 @@ export const updateTeam = functions.firestore.document('teams/{teamId}').onUpdat
 
     if ((oldName !== newName)
         || (oldAccessLevel !== newAccessLevel)
-        || (!mapArraysEquals(oldMembers, newMembers))) {
+        || (!utils.mapArraysEquals(oldMembers, newMembers))) {
         const data = {
             team: {
                 id: teamId,
@@ -278,52 +279,6 @@ export const updateUser = functions.firestore.document('users/{userId}').onUpdat
 
     await batch.commit();
 });
-
-function mapArraysEquals(array1: Array<Map<string, any>>, array2: Array<Map<string, any>>): boolean {
-    if (array1.length !== array2.length) {
-        return false;
-    }
-
-    const sortedArray1 = array1.sort(mapComparer);
-    const sortedArray2 = array2.sort(mapComparer);
-
-    for (let i = 0; i < sortedArray1.length; i++) {
-        if (!mapEquals(sortedArray1[i], sortedArray2[i]))
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-function mapEquals(map1: Map<string, any>, map2: Map<string, any>): boolean {
-    if (map1.size !== map2.size)
-    {
-        return false;
-    }
-
-    for (const key in map1.keys) {
-        if (!map2.has(key) || map1.get(key) !== map2.get(key))
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-function mapComparer(map1: Map<string, any>, map2: Map<string, any>) {
-    if (map1.get("id") > map2.get("id")) {
-        return 1;
-    }
-
-    if (map1.get("id") < map2.get("id")) {
-        return -1;
-    }
-
-    return 0;
-}
 
 class User {
     name: string;
